created testing for vendor product name
